+++ src/charactersData/characters/Raiden.ts
@@ -43,7 +43,7 @@ const Raiden = {
   }),
   possible_teams: [
     [Character.RAIDEN, Character.XINGQIU, Character.XIANGLING, Character.BENNETT],
-    [Character.RAIDEN, Character.SARA, Character.SUCROSE, Character.BENNETT],
+    [Character.RAIDEN, Character.KUJOU_SARA, Character.SUCROSE, Character.BENNETT],
     [Character.EULA, Character.RAIDEN, Character.ROSARIA, Character.BENNETT],
   ]
 }