+++ src/views/Character/components/Materials/index.tsx
@@ -2,40 +2,73 @@ import {FC, ReactElement, ReactNode} from 'react'
 import {RouteComponentProps, withRouter} from 'react-router-dom';
 import {useTranslation} from 'react-i18next';
 import {Col, Row} from 'react-bootstrap';
-import { v4 as uuid } from 'uuid';
+import {v4 as uuid} from 'uuid';
 
-import {ICharacter} from '../../../../charactersData/types';
+import {IAscensionMaterials, ICharacter} from '../../../../charactersData/types';
 import CHARACTERS from '../../../../charactersData';
 import Material from '../Material';
 import {IRouteParams} from '../../../../types/commonTypes';
 import CharacterImage from '../../../../common/CharacterImage';
 import Stars from '../../../../common/Stars';
 import AscensionSummary from '../AscensionSummary';
+import {Character} from '../../../../const/consts';
 
 const Materials: FC<RouteComponentProps<IRouteParams>> = ({match: {params}}): ReactElement => {
-  const {t} = useTranslation('character');
+  const {t} = useTranslation(['character', 'common']);
   const {name} = params
   const activeCharacter = CHARACTERS.find((character: ICharacter) => character.name === params.name)
 
   if (!activeCharacter) return <></>
 
   const ascensionMaterials = Object.entries(activeCharacter.ascension_materials)
-  const talentMaterials = activeCharacter?.talent_materials ?
-    Object.entries(activeCharacter.talent_materials) : []
 
   const {rarity} = activeCharacter
 
-  const renderTalentMaterials = (): ReactNode => talentMaterials.length > 0 && (
+  const renderMaterials = (talentInfo: IAscensionMaterials | undefined) => {
+    if (!talentInfo) return null
+
+    const materials = talentInfo ? Object.entries(talentInfo) : []
+
+    return (
+      <>
+        {materials.map((value, index) => {
+          if (!value[0]) return null
+
+          return <Material data={value} key={index}/>
+        })}
+        {talentInfo && <AscensionSummary ascensionMaterials={talentInfo}/>}
+      </>
+    )
+  }
+
+  const renderOtherCharactersTalentMaterials = (): ReactNode => activeCharacter?.talent_materials && (
     <Col xs={12}>
       <h4 className='mb-3'>{t('character:talents_enhancement')}</h4>
-      {talentMaterials.map((value, index) => {
-        if (!value[0]) return null
-
-        return <Material data={value} key={index}/>
-      })}
-      {activeCharacter?.talent_materials && <AscensionSummary ascensionMaterials={activeCharacter.talent_materials}/>}
+      {renderMaterials(activeCharacter?.talent_materials)}
     </Col>
   )
+
+  const renderTravelerTalentMaterials = (): ReactElement => (
+    <Col xs={12}>
+      <h4 className='mb-3'>{t('character:talents_enhancement')}</h4>
+
+      <h5 className='my-3'>{t('common:attack')}</h5>
+      {renderMaterials(activeCharacter.attack)}
+
+      <h5 className='my-4'>{t('common:elemental_skill')}</h5>
+      {renderMaterials(activeCharacter.elemental_skill)}
+
+      <h5 className='my-4'>{t('common:elemental_burst')}</h5>
+      {renderMaterials(activeCharacter.elemental_burst)}
+    </Col>
+  )
+
+  const renderTalentMaterials = (): ReactNode => {
+    if (activeCharacter.name === Character.TRAVELER) return renderTravelerTalentMaterials()
+
+    return renderOtherCharactersTalentMaterials()
+  }
+
 
   const renderPossibleTeams = (): ReactNode => activeCharacter.possible_teams && (
     <Col xs={12}>
@@ -59,7 +92,7 @@ const Materials: FC<RouteComponentProps<IRouteParams>> = ({match: {params}}): Re
       <Col sm={12} md={12} lg={8} xl={7} xxl={6}>
         <Row className='justify-content-center gx-3 gy-3'>
           <Col xs={12}>
-            <CharacterImage name={name}/>
+            <CharacterImage name={name} withBorder={true}/>
             <h1 className='fs-3 mb-0'>{t(`character:names.${name}`)}</h1>
             <Stars count={rarity}/>
           </Col>
@@ -77,6 +110,7 @@ const Materials: FC<RouteComponentProps<IRouteParams>> = ({match: {params}}): Re
           {renderTalentMaterials()}
 
           {renderPossibleTeams()}
+
         </Row>
       </Col>
     </Row>